move model_attribute validation to models that have model_attribute attribute
