+++ src/holon/models/rule_actions/rule_action_utils.py
@@ -8,6 +8,7 @@ from django.apps import apps
 from django.forms import ModelForm
 
 from polymorphic import utils
+from holon.models.scenario_rule import ScenarioRule
 from holon.models.util import all_subclasses
 
 
@@ -30,7 +31,9 @@ class RuleActionUtils:
         if base_class == Contract:
             return [(Actor, "owner_actor")]
 
-    def get_balanceable_subtypes():
+    def get_balanceable_subtypes() -> list[str]:
+        """Get all possible subclass types for the balanceable models"""
+
         base_classes = [EnergyAsset, GridConnection, Contract]
         choices = []
 
@@ -41,3 +44,11 @@ class RuleActionUtils:
             ]
 
         return choices
+
+    def get_model_attributes_options(rule: ScenarioRule) -> list[str]:
+        """Get the possible asset attributes for a certain model type"""
+
+        model_type = rule.model_type if rule.model_subtype is None else rule.model_subtype
+        model = apps.get_model("holon", model_type)
+
+        return [field.name for field in model()._meta.get_fields() if not field.is_relation]
